refactor(cli): extract the REPL and one-shot commands from main.rs (#518)
* refactor(cli): extract the REPL and one-shot commands from main.rs

Move start_interpreter (REPL driver with history handling and brace
counting) to repl.rs and execute_script / execute_format /
report_diagnostics to commands.rs, leaving main.rs as clap wiring
only. The cast / invoke / align surface is unchanged.

Closes #509


* chore(codecov): ignore the abyss-cli shell crate

The CLI crate is argv dispatch, the interactive rustyline REPL, and
thin print wrappers over covered library calls; nothing in it is
reachable from cargo test, so the code move in this PR tripped the
patch target despite adding no logic.


---------
